audiodelays, audiofilters, audiofreeverb: buffer lengths and silence fills
Echo and MultiTapDelay applied their length floor after their ceiling, so a
short delay could be raised above the maximum and the memset that follows
took "max - len" as an unsigned length. Chorus never clamped its delay to the
buffer it allocated. PitchShift accepted a window too small to hold one
sample.

The unsigned 16-bit silence fill used memset with 32768, which repeats a
single byte and so wrote zeros -- full negative deflection rather than the
midpoint -- in PitchShift, GranularPitchShift and Distortion. Distortion's
hard clip had an upper bound of 32768, which becomes -32768 on the cast to
int16_t below it.

Freeverb cleared only half of each delay line, counting uint16_t entries as
bytes, and declared its channel offsets inside the per-sample loop, so both
channels went through the left bank of filters. The biquad chain counted a
size_t length with a uint8_t in four places.

+++ shared-module/audiofilters/Filter.c
@@ -225,7 +225,7 @@ audioio_get_buffer_result_t audiofilters_filter_get_buffer(audiofilters_filter_o
                     }
 
                     // Process biquad filters
-                    for (uint8_t j = 0; j < self->filter.objs_len; j++) {
+                    for (size_t j = 0; j < self->filter.objs_len; j++) {
                         mp_obj_t filter_obj = self->filter.objs[j];
                         common_hal_synthio_biquad_tick(filter_obj);
                         for (uint8_t k = 0; k < self->base.channel_count; k++) {
